changed transaction to two-stage struct

+++ src/hotstuff/client_command.rs
@@ -1,13 +1,14 @@
-use crate::types::Sha256Hash;
 use serde::{Deserialize, Serialize};
 use sha2::{Digest, Sha256};
+
+use crate::types::transaction::Sha256Hash;
 
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
 pub enum Action {
     Transfer {
         from: String,
         to: String,
-        amount: u64,
+        amount: u128,
     },
     Empty,
 }
